Bring CVS tree in sync with the new 2.0.1 stable
Fossil-ID: SVN r6

+++ clients/upslog.c
@@ -153,7 +153,7 @@ static void do_upshost(const char *arg)
 
 static void do_pid(const char *arg)
 {
-	snprintfcat(logbuffer, sizeof(logbuffer), "%d", getpid());
+	snprintfcat(logbuffer, sizeof(logbuffer), "%ld", (long)getpid());
 }
 
 static void do_time(const char *arg)
